Make renderer a bit more modular Switch back to float in waveForm_t

+++ code/renderer/tr_local.h
@@ -208,10 +208,10 @@ typedef enum {
 } acff_t;
 
 typedef struct {
-	double base;		// -EC- set to double for frameloss fix
-	double amplitude;
-	double phase;
-	double frequency;
+	float base;
+	float amplitude;
+	float phase;
+	float frequency;
 
 	genFunc_t	func;
 } waveForm_t;
